Added data for fine-tuning models; refactored primitives for fine-tuning models,
