+++ symai/ops/symbol_primitives.py
@@ -1077,3 +1077,24 @@ class OutputHandlingPrimitives:
 
         return self._to_symbol(_func(self, *args))
 
+
+
+class FineTuningPrimitives:
+    '''
+    This mixin contains functionalities related to fine tuning models.
+    '''
+    def tune(self, operation: str = 'create', **kwargs) -> 'Symbol':
+        '''
+        Fine tune a base model.
+
+        Args:
+            operation (str, optional): The specific operation to be performed. Defaults to 'create'.
+            **kwargs: Additional keyword arguments to be passed to the `@core.tune` decorator dependent on the used operation.
+
+        Returns:
+            Symbol: The resulting Symbol containing the fine tuned model ID.
+        '''
+        @core.tune(operation=operation, **kwargs)
+        def _func(_, *args, **kwargs) -> str:
+            pass
+        return self.sym_return_type(_func(self))
